requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
build==1.2.2.post1
    # via pip-tools
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via pip-tools
execnet==2.1.1
    # via pytest-xdist
greenlet==3.0.1
    # via playwright
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via pytest-html
markupsafe==3.0.2
    # via jinja2
packaging==25.0
    # via
    #   build
    #   pytest
    #   pytest-rerunfailures
pip-tools==7.4.1
    # via -r requirements.in
playwright==1.40.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.6.0
    # via pytest
pydantic==2.11.7
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pyee==11.0.1
    # via playwright
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==7.4.3
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-base-url
    #   pytest-html
    #   pytest-metadata
    #   pytest-playwright
    #   pytest-rerunfailures
    #   pytest-xdist
pytest-asyncio==0.21.1
    # via -r requirements.in
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-html==4.1.1
    # via -r requirements.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-playwright==0.4.3
    # via -r requirements.in
pytest-rerunfailures==15.1
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.4
    # via pytest-base-url
sniffio==1.3.1
    # via anyio
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.14.1
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
